split owners-aliases-bumper in smaller files
